Implemented discussion selection
A bit tricky because we have to make two different controllers
communicate.

The implementation uses a service injected by Angular to both
controllers. The receiver (here, DiscussController) watches a
modification on the value contained by the service. When the value is
modified, it updates its scope accordingly, thus modifying the
discussion shown.